fix: 7-Agent Cross Audit P3 21건 — 보안/성능/UI 개선
Security: Field(max_length) 입력 검증, SQL injection 방지, StaticFiles html=False
Performance: 정규식 프리컴파일, LRU 캐시(256), IGNORE_TOKENS 라이브 함수
Stability: seed BigInteger, LoRA source DB 캐시 기반, SERVER_HOST/PORT 환경변수화
UI/UX: 반응형 그리드, CLS 방지, Tooltip delay, sceneTextFont 용어 통일

+++ backend/models/prompt_history.py
@@ -1,6 +1,6 @@
 """Prompt History model for V3 Schema."""
 
-from sqlalchemy import Boolean, Float, Integer, String, Text
+from sqlalchemy import BigInteger, Boolean, Float, Integer, String, Text
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -24,7 +24,7 @@ class PromptHistory(Base, TimestampMixin, SoftDeleteMixin):
     steps: Mapped[int] = mapped_column(Integer, default=20)
     cfg_scale: Mapped[float] = mapped_column(Float, default=7.0)
     sampler_name: Mapped[str | None] = mapped_column(String(50))
-    seed: Mapped[int | None] = mapped_column(Integer)
+    seed: Mapped[int | None] = mapped_column(BigInteger)
     clip_skip: Mapped[int] = mapped_column(Integer, default=2)
 
     # Bundles
